requirements.txt
langchain
langchain_core
langchain_community
langchain_google_genai
langchain_huggingface
langchain_chroma
python-dotenv
pydantic
sentence-transformers
pypdf
